Add files via upload
# Data Exploration

## Dataset

The Dataset I will be wrangling, analyzing and communicate my findings is the Prosper Loan Dataset. This data set is a record of loans given out by Prosper, a loan company in the United States. (www.prosper.com) The loan data contains records of 113,937 loans in total. 

The purpose of this analysis is to investigate the performance of loans by looking at the dataset in terms of completed repayments, profits and return on investments and looking at various trends and factors which may be responsible for performance of loans. I would summarize in this report, the actions taken to analyze the various relationships: univariate, bivariate, and multivariate between variables in the dataset.

The features of interest to this study are loan status, Employment Status, the borrower's home ownership status, state, verifiable income, and occupation. Steps were taken to clean the data to make it suitable for analysis. These steps include filling null values, dropping duplicate rows, renaming columns and transforming names.

This is an Exploratory and Explanatory Data Analaysis where I will be using visualizations to communicate my findings.


## Summary of Findings

During exploration, I found a strong connection between the borrower annual percentage rate (apr) and the borrower rate.

I discovered that majority of the borrowers have a Prosper Score of 4.

Also, That About 50% of the total loans are current, (ongoing) 34% are completed, 10.7% are Chargedoff while 4.5% are defaulted.

That when the borrower has a good prosper score , this has a negative effect on the lendersyield.

That Current loans seems to have better prosper score than other loan status.There is a slight difference between the estimates of yields and return. 

There is no significance whether a borrower owns a home or does not.

#### I will bring these findings into my explanatory presentation.


## Key Insights for Presentation

Introducing the dataset and important variables, then show the pattern in distribution, then plot the scatterplot.

Presentation also looks at whether home ownership determines a borrower's ability to settle his/her loans.